Added article to cover the use of %nuget% - Issue #1
Added an article to cover the use of the %nuget% environment variable within a build script.  This should complete Issue #1.
I have stubbed out what I have called a Build Services Reference page.  The idea would be to expand on this to cover all the Build Services functionality, and to then include all the other services that are offered by MyGet.