Resolves #21 - Added how to for adding new feed owner
This how to describes how to add a new feed owner.  It links to a new page called the Feed Security Reference Page.  This page will need some more work in order to complete all necessary information.  This simply lists the available permissions that can be assigned to a feed.  I will create a new issue for compeleting this page.